refactor: 댓글 및 게시물 관련 폴더 이름름 수정
