+++ src/entities/comment/actions/useDeleteComment.ts
