+++ src/entities/post/model/index.ts
@@ -1,0 +1,12 @@
+/**
+ * 게시물 관련 모델과 쿼리 훅 내보내기
+ */
+
+// 쿼리 훅 내보내기
+export { usePostsQuery } from './postsQuery'
+export { usePostsByTagQuery } from './postsByTagQuery'
+export { useSearchPostsQuery } from './searchPostsQuery'
+export { useTagsQuery } from './tagsQuery'
+
+// 타입 내보내기
+export type { PostsQueryParams } from './postsQuery'
