+++ src/features/comment/ui/AddCommentDialog.tsx
@@ -1,7 +1,7 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle, Textarea, Button } from "../../../shared/ui"
 import useCommentStore from "../model/useCommentStore"
 import { useShallow } from "zustand/shallow"
-import { useCreateComment } from "../../../entities/comment/hooks/useCreateComment"
+import { useCreateComment } from "../../../entities/comment/actions/useCreateComment"
 
 const AddCommentDialog = () => {
   const { showAddCommentDialog, setShowAddCommentDialog, newComment, setNewComment } = useCommentStore(
