+++ src/features/posts/api/index.ts
@@ -1,2 +1,2 @@
-export * from "../../../entities/post/hooks/useDeletePost"
-export * from "../../../entities/post/hooks/useUpdatePost"
+export * from "../../../entities/post/actions/useDeletePost"
+export * from "../../../entities/post/actions/useUpdatePost"
